pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "graphcap"
version = "0.0.1"
description = "graphcap"
requires-python = ">=3.11"
dependencies = [
    "click>=8.1.8",
    "datasets>=3.2.0",
    "fastapi>=0.115.6",
    "httpx>=0.28.1",
    "huggingface-hub>=0.27.1",
    "jinja2>=3.1.5",
    "loguru>=0.7.3",
    "matplotlib>=3.10.0",
    "networkx>=3.4.2",
    "openai>=1.59.9",
    "pydantic>=2.10.5",
    "pytest-watcher>=0.4.3",
    "python-dotenv>=1.0.1",
    "python-multipart>=0.0.20",
    "pyvis>=0.3.2",
    "rich>=13.9.4",
    "scipy>=1.15.1",
    "tenacity>=9.0.0",
    "termcolor>=2.5.0",
    "uvicorn>=0.34.0",
]

# Add this section to specify where your package source code is located
[tool.hatch.build.targets.wheel]
packages = [
    "graphcap",
] # Adjust this path to match your actual source code location

[tool.uv]
dev-dependencies = [
    "basedpyright>=1.26.0",
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "httpx>=0.24.0",
    "pytest-cov>=4.1.0",
]

[project.scripts]
graphcap = "graphcap.main:cli"

[project.optional-dependencies]
dev = []

# Added Ruff configuration
[tool.ruff]
# Exclude a variety of commonly ignored directories
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
]

# Source directories
src = ["./graphcap", "./server"]

# Same as Black
line-length = 120
indent-width = 4

# Assume Python 3.11
target-version = "py311"

[tool.ruff.lint]
# Enable Pyflakes (`F`) and a subset of the pycodestyle (`E`) codes by default
select = ["E", "F", "W", "I001"]
ignore = []

# Allow fix for all enabled rules (when `--fix`) is provided
fixable = ["ALL"]
unfixable = []

# Allow unused variables when underscore-prefixed
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
